requirement.txt
jupyter==1.0.0
numpy==1.15.2
pandas==0.23.4
scipy==1.1.0
statsmodels==0.9.0
scikit-learn==0.20.3
joblib==0.13.2
matplotlib==3.0.3
requests==2.21.0
pyramid-arima==0.8.1